# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: HardCoreVisor API
  description: >
    REST API for HardCoreVisor — a hybrid virtualisation platform combining
    Rust (KVM core) and Go (orchestration controller) with triple VMM backends
    (rust-vmm for high-performance Linux microVMs, QEMU for general-purpose VMs,
    LXC for lightweight containers).
  version: "0.1.0"
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:18080
    description: Local development controller
...
paths:
...
  /api/v1/auth/users:
    get:
      tags: [Auth]
      summary: List users (admin only)
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserInfo"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Auth]
      summary: Create a user (admin only)
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/auth/users/{username}:
    delete:
      tags: [Auth]
      summary: Delete a user (admin only)
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
  schemas:
...
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          description: URI reference identifying the problem type
          example: "about:blank"
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation
        instance:
          type: string
          description: URI reference for this occurrence
        error:
          type: string
          description: Legacy error field (backward compat)

    # ── Common ─────────────────────────────────────────
...
    UserInfo:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
        created_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        role:
          type: string
          enum: [admin, operator, viewer]

    # ── Monitoring ─────────────────────────────────────
...